Consolidate date adapter
